+++ gulpfile.js
@@ -13,6 +13,7 @@ const del = require('del');
 const through = require('through2');
 const svgr = require('@svgr/core').default;
 const babelCore = require('@babel/core');
+const cheerio = require('cheerio');
 
 const sep = path.sep;
 const dest = gulp.dest;
@@ -89,6 +90,7 @@ gulp.task('clean:index', () => del('src/index.ts'));
 gulp.task('svg', () => {
     return gulp.src('src/**/*.svg')
         .pipe(svgmin(file => getSvgrConfig(file.path).svgoConfig))
+        .pipe(removeSvgMasks())
         .pipe(svgToReact())
         .pipe(dest(dist));
 });
@@ -261,3 +263,90 @@ const getComponentName = filePath => {
         .map(w => w.charAt(0).toUpperCase() + w.slice(1))
         .join('');
 };
+
+
+const removeSvgMasks = () => changePipe(function (file, encoding) {
+    const content = file.contents.toString(encoding);
+    const $ = cheerio.load(content, { xmlMode: true });
+    const $svg = $('svg');
+
+    const mask = $svg.find('[mask]');
+
+    if (mask.length === 0) {
+        this.push(file);
+        return;
+    }
+
+    mask.each((i, item) => getMaskData(item));
+
+    file.contents = Buffer.from($.xml());
+
+    this.push(file);
+
+    function getMaskData(mask) {
+        const maskAttr = $(mask).attr('mask');
+        const maskId = getIdFromMaskAttr(maskAttr);
+        const divs = [];
+
+        if (maskId.length === 0) {
+            return;
+        }
+
+        findById(maskId, divs);
+
+        if (divs.length === 0) {
+            return;
+        }
+
+        $(mask).html(divs.map(i => i.removeAttr('id')).join(''));
+
+        $(mask).removeAttr('mask')
+    }
+
+    function linkedTag($tag, divs, $parent) {
+        const attrs = $tag.attr();
+        const { 'xlink:href': xlinkHref } = attrs;
+
+        switch (true) {
+            case !!xlinkHref:
+                findById(xlinkHref, divs, $parent);
+                break;
+            default:
+                divs.push($tag);
+                $tag.remove();
+                $parent.remove();
+                break;
+        }
+    }
+
+    function findById(id, divs, $parent) {
+        const tag = $svg.find(id);
+
+        if (tag.length === 0) {
+            return;
+        }
+
+        const child = tag.children();
+
+        switch (child.length) {
+            case 1:
+                linkedTag(child, divs, tag);
+                break;
+            case 0:
+                divs.push(tag);
+                tag.remove();
+                $parent && $parent.remove();
+                break;
+            default:
+                child.map(item => linkedTag(item, divs));
+                break;
+        }
+    }
+
+    function getIdFromMaskAttr(attr) {
+        return attr.match(/\(.*\)/g).reduce(
+            (id, match) => match ? match.slice(1, -1) : id,
+            ''
+        );
+    }
+});